feat(#60): moved FederalCurriculumRequirements to requirements package
