+++ src/main/java/com/eshabakhov/schoodule/federal/curriculum/FcFull.java
@@ -5,7 +5,6 @@ package com.eshabakhov.schoodule.federal.curriculum;
 
 import com.eshabakhov.schoodule.Media;
 import com.eshabakhov.schoodule.federal.FederalCurriculum;
-import com.eshabakhov.schoodule.federal.FederalCurriculumRequirements;
 
 /**
  * Full implementation of {@link FederalCurriculum}.
