+++ src/main/java/com/eshabakhov/schoodule/federal/curriculum/FederalCurriculumRequirement.java
@@ -1,7 +1,7 @@
 /*
  * © 2025-2026 Eset Shabakhov. Schoodule
  */
-package com.eshabakhov.schoodule.federal;
+package com.eshabakhov.schoodule.federal.curriculum;
 
 import com.eshabakhov.schoodule.Media;
 import com.eshabakhov.schoodule.Printable;
