+++ src/main/java/com/eshabakhov/schoodule/federal/curriculum/requirement/FcrPostgres.java
@@ -5,7 +5,7 @@ package com.eshabakhov.schoodule.federal.curriculum.requirement;
 
 import com.eshabakhov.schoodule.Media;
 import com.eshabakhov.schoodule.enums.CurriculumPartType;
-import com.eshabakhov.schoodule.federal.FederalCurriculumRequirement;
+import com.eshabakhov.schoodule.federal.curriculum.FederalCurriculumRequirement;
 import org.jooq.DSLContext;
 
 /**
