+++ src/main/java/com/eshabakhov/schoodule/federal/curriculum/requirement/FcrsPostgres.java
@@ -6,8 +6,8 @@ package com.eshabakhov.schoodule.federal.curriculum.requirement;
 import com.eshabakhov.schoodule.Page;
 import com.eshabakhov.schoodule.PageableList;
 import com.eshabakhov.schoodule.enums.CurriculumPartType;
-import com.eshabakhov.schoodule.federal.FederalCurriculumRequirement;
-import com.eshabakhov.schoodule.federal.FederalCurriculumRequirements;
+import com.eshabakhov.schoodule.federal.curriculum.FederalCurriculumRequirement;
+import com.eshabakhov.schoodule.federal.curriculum.FederalCurriculumRequirements;
 import com.eshabakhov.schoodule.page.ResponsePageableList;
 import com.eshabakhov.schoodule.tables.records.FederalCurriculumRequirementRecord;
 import lombok.EqualsAndHashCode;
