+++ src/main/java/com/eshabakhov/schoodule/federal/curriculum/requirement/FederalCurriculumRequirementHtmlController.java
@@ -3,7 +3,7 @@
  */
 package com.eshabakhov.schoodule.federal.curriculum.requirement;
 
-import com.eshabakhov.schoodule.federal.FederalCurriculumRequirement;
+import com.eshabakhov.schoodule.federal.curriculum.FederalCurriculumRequirement;
 import com.eshabakhov.schoodule.federal.curriculum.FcsPostgres;
 import org.jooq.DSLContext;
 import org.springframework.security.access.prepost.PreAuthorize;
